+++ spec/features/memoization_spec.rb
@@ -12,6 +12,10 @@ RSpec.describe 'Memoization (dependency memoization)' do
     second_resolve = container['deps.sidekiq']
 
     expect(first_resolve.object_id).not_to eq(second_resolve.object_id)
+
+    # runtime non-memoized registration
+    container.register('no_memoized') { Object.new }
+    expect(container['no_memoized'].object_id).not_to eq(container['no_memoized'].object_id)
   end
 
   specify 'explicit memoization boolean flag (memoize or not)' do
@@ -41,5 +45,20 @@ RSpec.describe 'Memoization (dependency memoization)' do
     first_reveal  = container['nonmemoized.sneakers']
     second_reveal = container['nonmemoized.sneakers']
     expect(first_reveal.object_id).not_to eq(second_reveal.object_id)
+
+    # runtime non-memoized registration
+    container.register('non_memoized') { Object.new }
+    expect(container['non_memoized'].object_id).not_to eq(container['non_memoized'].object_id)
+
+    container.register('expl_non_memoized', memoize: false) { Object.new }
+    first_reveal  = container['expl_non_memoized']
+    second_reveal = container['expl_non_memoized']
+    expect(first_reveal.object_id).not_to eq(second_reveal.object_id)
+
+    # runtime memoized registration
+    container.register('expl_memoized', memoize: true) { Object.new }
+    first_reveal  = container['expl_memoized']
+    second_reveal = container['expl_memoized']
+    expect(first_reveal.object_id).to eq(second_reveal.object_id)
   end
 end